Moved pended claim updates to new class
